+++ src/CognitiveServices/CognitiveServices/CognitiveServicesAccount/RemoveAzureCognitiveServicesAccountDeployment.cs
@@ -1,0 +1,128 @@
+﻿// ----------------------------------------------------------------------------------
+//
+// Copyright Microsoft Corporation
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// http://www.apache.org/licenses/LICENSE-2.0
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+// ----------------------------------------------------------------------------------
+
+using Microsoft.Azure.Commands.Management.CognitiveServices.Properties;
+using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
+using Microsoft.Azure.Management.CognitiveServices;
+using Microsoft.Azure.Management.CognitiveServices.Models;
+using System;
+using System.Globalization;
+using System.Management.Automation;
+
+namespace Microsoft.Azure.Commands.Management.CognitiveServices
+{
+    /// <summary>
+    /// Delete a Cognitive Services.
+    /// </summary>
+    [Cmdlet("Remove", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "CognitiveServicesAccountDeployment", DefaultParameterSetName = DefaultParameterSet, SupportsShouldProcess = true), OutputType(typeof(bool))]
+    public class RemoveAzureCognitiveServicesAccountDeploymentCommand : CognitiveServicesAccountBaseCmdlet
+    {
+        protected const string DefaultParameterSet = "DefaultParameterSet";
+        protected const string ResourceIdParameterSet = "ResourceIdParameterSet";
+        protected const string InputObjectParameterSet = "InputObjectParameterSet";
+
+        [Parameter(
+            Position = 0,
+            Mandatory = true,
+            ParameterSetName = InputObjectParameterSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "InputObject.")]
+        [ValidateNotNullOrEmpty]
+        public Deployment InputObject { get; set; }
+
+        [Parameter(
+            Position = 0,
+            Mandatory = true,
+            ParameterSetName = ResourceIdParameterSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Resource Id.")]
+        [ValidateNotNullOrEmpty]
+        public string ResourceId { get; set; }
+
+        [Parameter(
+            Position = 0,
+            Mandatory = true,
+            ParameterSetName = DefaultParameterSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Resource Group Name.")]
+        [ResourceGroupCompleter()]
+        [ValidateNotNullOrEmpty]
+        public string ResourceGroupName { get; set; }
+
+        [Parameter(
+            Position = 1,
+            Mandatory = true,
+            ParameterSetName = DefaultParameterSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Cognitive Services Account Name.")]
+        [Alias(CognitiveServicesAccountNameAlias)]
+        [ValidateNotNullOrEmpty]
+        public string AccountName { get; set; }
+
+        [Parameter(
+            Position = 2,
+            Mandatory = true,
+            ParameterSetName = DefaultParameterSet,
+            ValueFromPipelineByPropertyName = true,
+            HelpMessage = "Cognitive Services CommitmentPlan Name.")]
+        [ValidateNotNullOrEmpty]
+        public string Name { get; set; }
+
+        [Parameter(Mandatory = false, HelpMessage = "Don't ask for confirmation.")]
+        public SwitchParameter Force { get; set; }
+
+        [Parameter(Mandatory = false)]
+        public SwitchParameter PassThru { get; set; }
+
+        public override void ExecuteCmdlet()
+        {
+            base.ExecuteCmdlet();
+
+            if (ShouldProcess(
+                this.AccountName, string.Format(CultureInfo.CurrentCulture, Resources.RemoveAccount_ProcessMessage, this.AccountName))
+                ||
+                Force.IsPresent)
+            {
+                RunCmdLet(() =>
+                {
+                    switch (ParameterSetName)
+                    {
+                        case InputObjectParameterSet:
+                            ResourceId = InputObject.Id;
+                            goto case ResourceIdParameterSet;
+                        case ResourceIdParameterSet:
+                            if (!CognitiveServices.ResourceId.TryParse(ResourceId, out CognitiveServices.ResourceId resourceId))
+                            {
+                                WriteError(new ErrorRecord(new Exception("Failed to parse ResourceId"), string.Empty, ErrorCategory.NotSpecified, null));
+                            }
+
+                            ResourceGroupName = resourceId.ResourceGroupName;
+                            AccountName = resourceId.GetAccountName();
+                            Name = resourceId.GetAccountSubResourceName();
+                            break;
+                        case DefaultParameterSet:
+                            break;
+                    }
+
+                    this.CognitiveServicesClient.Deployments.Delete(ResourceGroupName, AccountName, Name);
+
+                    if (PassThru.IsPresent)
+                    {
+                        WriteObject(true);
+                    }
+                });
+            }
+        }
+    }
+}