+++ src/Compute/Compute/ChangeLog.md
@@ -22,6 +22,7 @@
 ## Upcoming Release
 * Changed `New-AzVM` cmdlet when using the SimpleParameterSet to not create a PublicIPAddress when a PublicIPAddress name is not provided. 
 * Add `PlatformFaultDomain` parameter to cmdlets: `New-AzVM` and `New-AzVMConfig`
+* Added `-Feature` parameter for `New-AzGalleryImageDefinition`
 
 ## Version 4.21.0
 * Contains updates to the following powershell cmdlets